Refactor simObjects treatment and RB-related classes
