+++ spg/include/spg/sim/solver/baseSolver.h
@@ -16,12 +16,12 @@ class BaseSolver
 public:
     virtual ~BaseSolver() = default;
     virtual void step() = 0;
-    void addSimObject(const SimObject &object);
-    void addRigidBodyGroup(const RigidBodyGroup &rbGroup);
-    std::vector<SimObject> &simObjects() { return m_simObjects; }
-    const std::vector<SimObject> &simObjects() const { return m_simObjects; }
-    std::vector<RigidBodyGroup> &rbGroups() { return m_rigidBodyGroups; }
-    const std::vector<RigidBodyGroup> &rbGroups() const { return m_rigidBodyGroups; }
+    template <typename TObj>
+    void addObject(const TObj &object);
+    std::vector<SimObject> &simObjects() { return std::get<std::vector<SimObject>>(m_objects); }
+    const std::vector<SimObject> &simObjects() const { return std::get<std::vector<SimObject>>(m_objects); }
+    std::vector<RigidBodyGroup> &rbGroups() { return std::get<std::vector<RigidBodyGroup>>(m_objects); }
+    const std::vector<RigidBodyGroup> &rbGroups() const { return std::get<std::vector<RigidBodyGroup>>(m_objects); }
     void setDt(Real dt) { m_dtStep = dt; }
     Real dt() { return m_dtStep; }
     void setNumSubsteps(int nsubsteps) { m_nsubsteps = nsubsteps; }
@@ -31,8 +31,7 @@ public:
     void setGravity(const Vector3 &gravity) { m_gravity = gravity; }
 
 protected:
-    std::vector<SimObject> m_simObjects;
-    std::vector<RigidBodyGroup> m_rigidBodyGroups;
+    std::tuple<std::vector<SimObject>, std::vector<RigidBodyGroup>> m_objects;
     Vector3 m_gravity{0, -9.8, 0};
     Real m_time{0};
     Real m_dtStep{0.01};
