+++ spg/source/sim/solver/implicitEulerNewtonRobust.cpp
@@ -19,7 +19,7 @@ void ImplicitEulerNewtonRobust::step()
 
     // Compute total DOFs
     int accumulatedNDOF = 0;
-    for (const auto &object : m_simObjects) {
+    for (const auto &object : std::get<std::vector<SimObject>>(m_objects)) {
         accumulatedNDOF += object.nDOF();
     }
     const int totalNDOF{accumulatedNDOF};
